Adicionado validacoes no delete matricula

+++ matriculas/delMatrBD.php
@@ -3,14 +3,25 @@
 	include_once ("../funcoes.php");
 	include_once ("../conexao.php");
 
-	foreach ($_POST['sel'] as $codigo) {
-		$query = "DELETE FROM matricula WHERE id = $codigo";
-		$res = mysqli_query($con, $query); 
-		if ($res) {
-			header("Location: ../matriculas/controleMatriculas.php");
-		}else{
-			header("Location: ../matriculas/delMatr.php");
+
+	if (isset($_POST['sel'])) {
+		foreach ($_POST['sel'] as $codigo) {
+
+			$query = "DELETE FROM matricula WHERE id = $codigo";
+    		//unset($_SESSION['sel']);
+			$res = mysqli_query($con, $query); 
+			
+			if ($res) {
+    			$_SESSION['msn'] = "<div class='alert alert-success' role='alert'> Alterado com sucesso!</div>";
+				header("Location: ../matriculas/controleMatriculas.php");
+			}else{
+        		$_SESSION['msn'] = "<div class='alert alert-danger' role='alert'> Falha ao alterar!</div>";
+				header("Location: ../matriculas/deletaMat.php");
+			}
 		}
+	}else{
+		header("Location: ../matriculas/deletaMat.php");
 	}
 
+					
 ?>
